requirements.txt
python-dotenv
python-multipart
opencv-python
fastapi
bcrypt